Download Malawi administrative boundaries shapefiles from Humanitarian Data Exchange

Download Malawi administrative boundaries shapefiles from Humanitarian Data Exchange

download_boundaries_shapefiles(id, .unzip = TRUE)

Arguments

id

Character vector for unique identifier of administrative boundaries data for Malawi from Humanitarian Data Exchange

.unzip

Logical. Should zip file be unzipped and extracted? Default to TRUE

Value

If .unzip, list of shapefiles within the downloaded zip file else path to downloaded zip file. (Invisible) Shapefiles zip file downloaded to specified folder.

Examples

download_boundaries_shapefiles(id = "cod-ab-mwi")
#>                                                   Name   Length
#> 1             mwi_admbnda_adm0_nso_hotosm_20230405.cpg        5
#> 2             mwi_admbnda_adm0_nso_hotosm_20230405.dbf      421
#> 3             mwi_admbnda_adm0_nso_hotosm_20230405.prj      145
#> 4             mwi_admbnda_adm0_nso_hotosm_20230405.sbn      132
#> 5             mwi_admbnda_adm0_nso_hotosm_20230405.sbx      116
#> 6             mwi_admbnda_adm0_nso_hotosm_20230405.shp  2910420
#> 7         mwi_admbnda_adm0_nso_hotosm_20230405.shp.xml   240788
#> 8             mwi_admbnda_adm0_nso_hotosm_20230405.shx      108
#> 9             mwi_admbnda_adm1_nso_hotosm_20230405.cpg        5
#> 10            mwi_admbnda_adm1_nso_hotosm_20230405.dbf     1111
#> 11            mwi_admbnda_adm1_nso_hotosm_20230405.prj      145
#> 12            mwi_admbnda_adm1_nso_hotosm_20230405.sbn      148
#> 13            mwi_admbnda_adm1_nso_hotosm_20230405.sbx      116
#> 14            mwi_admbnda_adm1_nso_hotosm_20230405.shp  3306820
#> 15        mwi_admbnda_adm1_nso_hotosm_20230405.shp.xml   247506
#> 16            mwi_admbnda_adm1_nso_hotosm_20230405.shx      124
#> 17            mwi_admbnda_adm2_nso_hotosm_20230405.cpg        5
#> 18            mwi_admbnda_adm2_nso_hotosm_20230405.dbf    12002
#> 19            mwi_admbnda_adm2_nso_hotosm_20230405.prj      145
#> 20            mwi_admbnda_adm2_nso_hotosm_20230405.sbn      468
#> 21            mwi_admbnda_adm2_nso_hotosm_20230405.sbx      156
#> 22            mwi_admbnda_adm2_nso_hotosm_20230405.shp  7169888
#> 23        mwi_admbnda_adm2_nso_hotosm_20230405.shp.xml   305153
#> 24            mwi_admbnda_adm2_nso_hotosm_20230405.shx      356
#> 25            mwi_admbnda_adm3_nso_hotosm_20230405.cpg        5
#> 26            mwi_admbnda_adm3_nso_hotosm_20230405.dbf   222611
#> 27            mwi_admbnda_adm3_nso_hotosm_20230405.prj      145
#> 28            mwi_admbnda_adm3_nso_hotosm_20230405.sbn     4340
#> 29            mwi_admbnda_adm3_nso_hotosm_20230405.sbx      388
#> 30            mwi_admbnda_adm3_nso_hotosm_20230405.shp 17720128
#> 31        mwi_admbnda_adm3_nso_hotosm_20230405.shp.xml   306076
#> 32            mwi_admbnda_adm3_nso_hotosm_20230405.shx     3564
#> 33     mwi_admbndl_admALL_nso_hotosm_itos_20230405.cpg        5
#> 34     mwi_admbndl_admALL_nso_hotosm_itos_20230405.dbf    85993
#> 35     mwi_admbndl_admALL_nso_hotosm_itos_20230405.prj      145
#> 36     mwi_admbndl_admALL_nso_hotosm_itos_20230405.sbn    16980
#> 37     mwi_admbndl_admALL_nso_hotosm_itos_20230405.sbx     1012
#> 38     mwi_admbndl_admALL_nso_hotosm_itos_20230405.shp 10579884
#> 39 mwi_admbndl_admALL_nso_hotosm_itos_20230405.shp.xml   414308
#> 40     mwi_admbndl_admALL_nso_hotosm_itos_20230405.shx    14108
#> 41     mwi_admbndp_admALL_nso_hotosm_itos_20230405.cpg        5
#> 42     mwi_admbndp_admALL_nso_hotosm_itos_20230405.dbf   210455
#> 43     mwi_admbndp_admALL_nso_hotosm_itos_20230405.prj      145
#> 44     mwi_admbndp_admALL_nso_hotosm_itos_20230405.sbn     4300
#> 45     mwi_admbndp_admALL_nso_hotosm_itos_20230405.sbx      348
#> 46     mwi_admbndp_admALL_nso_hotosm_itos_20230405.shp    12224
#> 47 mwi_admbndp_admALL_nso_hotosm_itos_20230405.shp.xml   319825
#> 48     mwi_admbndp_admALL_nso_hotosm_itos_20230405.shx     3564
#> 49                     mwi_admbndt_adminUnitLookup.cpg        5
#> 50                     mwi_admbndt_adminUnitLookup.dbf      746
#> 51                 mwi_admbndt_adminUnitLookup.dbf.xml     7866
#>                   Date
#> 1  2023-04-10 13:36:00
#> 2  2023-04-10 13:36:00
#> 3  2023-04-10 13:36:00
#> 4  2023-04-10 13:36:00
#> 5  2023-04-10 13:36:00
#> 6  2023-04-10 13:36:00
#> 7  2023-04-10 13:47:00
#> 8  2023-04-10 13:36:00
#> 9  2023-04-10 13:38:00
#> 10 2023-04-10 13:38:00
#> 11 2023-04-10 13:38:00
#> 12 2023-04-10 13:38:00
#> 13 2023-04-10 13:38:00
#> 14 2023-04-10 13:38:00
#> 15 2023-04-10 13:47:00
#> 16 2023-04-10 13:38:00
#> 17 2023-04-10 13:40:00
#> 18 2023-04-10 13:40:00
#> 19 2023-04-10 13:40:00
#> 20 2023-04-10 13:40:00
#> 21 2023-04-10 13:40:00
#> 22 2023-04-10 13:40:00
#> 23 2023-04-10 13:47:00
#> 24 2023-04-10 13:40:00
#> 25 2023-04-10 13:41:00
#> 26 2023-04-10 13:41:00
#> 27 2023-04-10 13:41:00
#> 28 2023-04-10 13:41:00
#> 29 2023-04-10 13:41:00
#> 30 2023-04-10 13:41:00
#> 31 2023-04-10 13:47:00
#> 32 2023-04-10 13:41:00
#> 33 2023-04-10 13:44:00
#> 34 2023-04-10 13:44:00
#> 35 2023-04-10 13:44:00
#> 36 2023-04-10 13:44:00
#> 37 2023-04-10 13:44:00
#> 38 2023-04-10 13:44:00
#> 39 2023-04-10 13:47:00
#> 40 2023-04-10 13:44:00
#> 41 2023-04-10 13:46:00
#> 42 2023-04-10 13:46:00
#> 43 2023-04-10 13:46:00
#> 44 2023-04-10 13:46:00
#> 45 2023-04-10 13:46:00
#> 46 2023-04-10 13:46:00
#> 47 2023-04-10 13:47:00
#> 48 2023-04-10 13:46:00
#> 49 2023-04-07 11:35:00
#> 50 2023-04-07 11:35:00
#> 51 2023-04-07 11:35:00